Added DTO support

+++ CrashCourse.PetShop.UI.WebApi/Controllers/PetsController.cs
@@ -1,13 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using System.Linq;
 using CrashCourse.PetShop.Core.IServices;
 using CrashCourse.PetShop.Core.Models;
+using CrashCourse.PetShop.UI.WebApi.Dtos.Pets;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CrashCourse.PetShop.UI.WebApi.Controllers
 {
     [ApiController]
-    [Route("[controller]")]
+    [Route("api/[controller]")]
     public class PetsController : ControllerBase
     {
         private readonly IPetService _petService;
@@ -19,58 +20,80 @@ namespace CrashCourse.PetShop.UI.WebApi.Controllers
             _petTypeService = petTypeService;
         }
 
-        // POST
+        // POST api/Pets
         [HttpPost]
-        public ActionResult<Pet> Create(Pet pet)
+        public ActionResult<Pet> Create(PostPetDto pet)
         {
             if (pet == null)
                 return BadRequest("Pet cannot be null");
+            
+            var petType = _petTypeService.SavePetType(_petTypeService.NewPetType(pet.Type.ToLower()));
 
-            var petCreated = _petService.CreateAndSavePet(pet.Name, pet.Type, pet.BirthDate, pet.SoldDate, pet.Color, pet.Price);
+            var petCreated = _petService.CreateAndSavePet(pet.Name, petType, pet.BirthDate, pet.SoldDate, pet.Color, pet.Price);
             return Created("Pet created...", petCreated);
         }
         
-        // GET
+        // GET api/Pets
         [HttpGet]
-        public ActionResult<List<Pet>> GetALl()
+        public ActionResult<List<GetPetDto>> GetALl()
         {
             if (_petService.GetAllPets().Count == 0)
-                return NotFound();
+                return NotFound("No pets found");
             
-            return _petService.GetAllPets();
+            var pets = _petService.GetAllPets();
+            
+            var petsDto = pets.Select(pet => new GetPetDto
+                {
+                    Name = pet.Name,
+                    BirthDate = pet.BirthDate,
+                    SoldDate = pet.SoldDate,
+                    Color = pet.Color,
+                    Price = pet.Price,
+                    Type = pet.Type.Name
+                })
+                .ToList();
+
+            return petsDto;
         }
 
-        // GET 5 Cheapest
+        // GET api/Pets/Cheapest
         [HttpGet]
         [Route("Cheapest")]
         public ActionResult<List<Pet>> GetFiveCheapest()
         {
             if (_petService.GetFiveCheapestPets().Count == 0)
-                return NotFound();
+                return NotFound("No pets found");
             
             return _petService.GetFiveCheapestPets();
         }
 
-        [HttpPut("{id}")]
-        public ActionResult<Pet> Update(int id, Pet pet)
+        // PUT api/Pets/{id}
+        [HttpPut("{id:int}")]
+        public ActionResult<Pet> Update(int id, PutPetDto petDto)
         {
-            if (id != pet.Id)
-                return BadRequest("Id to change does not match pet id");
+            if (id == 0 || _petService.GetPetById(id) == null)
+                return NotFound("Pet does not exist...");
 
-            var petUpdate = _petService.UpdatePet(pet);
-
-            if (petUpdate != null)
-                return Ok(pet);
-
-            return BadRequest("Cannot update pet");
+            var pet = new Pet
+            {
+                Id = id,
+                BirthDate = petDto.BirthDate,
+                Color = petDto.Color,
+                Name = petDto.Name,
+                Price = petDto.Price,
+                SoldDate = petDto.SoldDate,
+                Type = _petTypeService.NewPetType(petDto.Type)
+            };
+            
+            return Accepted("Pet updated...", _petService.UpdatePet(pet));
         }
         
-        // DELETE
+        // DELETE api/Pets
         [HttpDelete]
         public ActionResult<Pet> Delete(int id)
         {
-            if (_petService.GetPetById(id) == null)
-                return NotFound();
+            if (id == 0 || _petService.GetPetById(id) == null)
+                return NotFound("Pet by the given id does not exist");
 
             _petService.DeletePet(id);
